+++ R/properties-table.R
@@ -1,5 +1,5 @@
 
-#' Set the table's position with respect to surrounding content.
+#' Set the table's position with respect to surrounding content
 #'
 #' Table position may be "left", "right" or "center". If you want text to
 #' wrap around the table, use "wrapleft" or "wrapright".
@@ -180,7 +180,7 @@ make_getter_setters("tabular_environment", "table", check_fun = is.string)
 #' Within knitr, huxtable labels will default to the same as the knitr chunk label.
 #' To turn off this behaviour, set `options(huxtable.autolabel = FALSE)`.
 #'
-#' If you use \href{http://bookdown.org}{bookdown}, and set a label on your
+#' If you use \href{https://bookdown.org}{bookdown}, and set a label on your
 #' table, the table [caption()] will automatically be prefixed with `(#label)`.
 #' You can then refer to the table using `@ref(label)`. `label` needs to start
 #' with `"tab:"`; if it doesn't, the `"tab:"` prefix will be added